refactoring project structure. adding string calculator kata
